+++ packages/windmill/src/services/reports/transmission.rs
@@ -39,7 +39,7 @@ pub struct UserData {
     pub precinct_code: String,
     pub registered_voters: i64,
     pub ballots_counted: i64,
-    pub voters_turnout: i64,
+    pub voters_turnout: f64,
     pub sboc_code: String,
     pub sboc_transmitted: String,
     pub sboc_date_transmitted: String,
